Создание App + title
